ddr: timing:gxl p281 ddr set 2048 compile fail [1/1]
PD#OTT-7152

Problem:
define CONFIG_DDR_SIZE 2048
compile fail

Solution:
define CONFIG_DDR_TOTAL_SIZE (((unsigned int)CONFIG_DDR_SIZE) << 20)

Verify:
p281 test pass.

+++ board/amlogic/gxl_p281_v1/firmware/ramdump.h
@@ -44,7 +44,7 @@
  *                            |        |
  *     COMPRESS_START_ADDR -> +--------+
  */
-#define CONFIG_DDR_TOTAL_SIZE		(CONFIG_DDR_SIZE << 20)
+#define CONFIG_DDR_TOTAL_SIZE		(((unsigned int)CONFIG_DDR_SIZE) << 20)
 #define CONFIG_COMPRESSED_DATA_ADDR	(0x08000000)
 #define CONFIG_COMPRESSED_DATA_ADDR1	(0x08000000)
